adds pre-processor
