+++ Makefile
@@ -61,4 +61,8 @@ deploy:
 	cd $(WORKING_DIR)/terraform/workflow-manager; docker buildx build --platform linux/amd64 --progress=plain -t pennsieve/workflow-manager .
 	docker tag pennsieve/workflow-manager ${WM_REPO}
 	docker push ${WM_REPO}
+	@echo "Deploying pre-processor"
+	cd $(WORKING_DIR)/terraform/pre-processor; docker buildx build --platform linux/amd64 --progress=plain -t pennsieve/pre-processor .
+	docker tag pennsieve/pre-processor ${PRE_PROCESSOR_REPO}
+	docker push ${PRE_PROCESSOR_REPO}
 	cd $(WORKING_DIR) ; git clean -f ; git checkout -- .
