+++ terraform/application-gateway-lambda/application-gateway.py
@@ -16,7 +16,9 @@ def lambda_handler(event, context):
     container_name = os.environ['CONTAINER_NAME']
     security_group = os.environ['SECURITY_GROUP_ID']
     subnet_ids = os.environ['SUBNET_IDS']
+    task_definition_name_pre = os.environ['TASK_DEFINITION_NAME_PRE']
     task_definition_name_post = os.environ['TASK_DEFINITION_NAME_POST']
+    container_name_pre = os.environ['CONTAINER_NAME_PRE']
     container_name_post = os.environ['CONTAINER_NAME_POST']
     pennsieve_host = os.environ['PENNSIEVE_API_HOST']
     pennsieve_host2 = os.environ['PENNSIEVE_API_HOST2']
