+++ terraform/pre-processor/Readme.md
@@ -1,0 +1,11 @@
+To build:
+
+`docker build -t pennsieve/pre-processor .`
+
+On arm64 architectures:
+
+`docker build -f Dockerfile_arm64 -t pennsieve/pre-processor .`
+
+To run:
+
+`docker-compose up --build`
