+++ terraform/pre-processor/dev.env.example
@@ -1,0 +1,6 @@
+BASE_DIR=/service/data
+SESSION_TOKEN=<token>
+PENNSIEVE_API_HOST=https://api.pennsieve.net
+PENNSIEVE_API_HOST2=https://api2.pennsieve.net
+INTEGRATION_ID=<integrationId>
+PENNSIEVE_API_HOST=https://api2.pennsieve.net
