+++ terraform/pre-processor/docker-compose.yml
@@ -1,0 +1,15 @@
+version: '3.9'
+
+services:
+  
+  pre-processor:
+    env_file:
+      - dev.env
+    image: pennsieve/pre-processor
+    volumes:
+      - ./data:/service/data
+    container_name: pre-processor
+    build:
+      context: .
+      dockerfile: ./Dockerfile_arm64 # change to Dockerfile on Linux
+    entrypoint: /service/main
